oveanet-pack/seo-audit-360/extensions/google-analytics/requirements.txt
# Google Analytics 4 Extension — Dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-analytics-data>=0.18.0
